pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Marfa Nikitina <marfik93@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1.2"
django-bootstrap4 = "^22.2"
django-extensions="^3.2.1"
python-dotenv = "^0.21.0"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]
django = "4.1.2"
flake8 = "^4.0.1"
coverage = "^6.5.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
